feat(gui-v2): replace old index with list view
